refactor(sample): remove invalid file
remove invalid files
