--- samples/base/_TestScript.ts
@@ -1,7 +1,0 @@
-import { ComponentBase } from "../../engine/components/ComponentBase";
-
-export class TestScript extends ComponentBase {
-
-    
- 
-}
